Buy process progress, refactoring and quality of life changes
Buy now redirects to login when not logged in. Did really big refactoring and changed design layout from slide-bar to "reloading the page". It's much more convenient now. Made a show comments button that will show the individual post and its comments on one page where you can also make new comments and stuff and made a general routing path with a menu on the starting page
