--- ese2021-project-scaffolding/scaffolding/frontend/src/app/IEvent.ts
@@ -1,8 +1,0 @@
-/**
- * Event interface for post and comment
- */
-
-export interface IEvent {
-    delete(): void;
-    update(): void;
-}
