+++ ese2021-project-scaffolding/scaffolding/frontend/src/app/post/post.component.html
@@ -1,107 +1,68 @@
-<!--This is how the posts will appear on the front page-->
-<p class="post-information">
-  {{post.category}}
-  {{post.createdAt |  date:'dd/MM/yyyy'}}
-</p>
+<!-- // INTERACTION BUTTONS \\ -->
+<mat-card class="forum-menu" *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()">
+  <button class="post-button" mat-raised-button (click)="openPostWindow()">New Post</button>
 
-<mat-card class="title">
-  {{post.title}}
+  <mat-form-field>
+    <mat-label class="select-category">Filter</mat-label>
+    <mat-select [(ngModel)]="selectedTags" [formControl]="tags" multiple>
+      <mat-option *ngFor="let tag of tagList" [value]="tag">{{tag}}</mat-option>
+      <button mat-raised-button (click)="filterPosts()" >Apply</button>
+    </mat-select>
+  </mat-form-field>
 </mat-card>
+<!--/////////////////////////////////////////////////////////////////////////////////////-->
 
-<mat-card class="text">
-  {{post.text}}
-</mat-card>
-  
-<!--<img mat-card-image class="image" alt="image"
-  src="http://localhost:3000/src/public/uploads/1637796357453_Unknown.png">-->
+<!-- // DISPLAY POSTS \\ -->
+<mat-list-item *ngFor="let post of posts">
+  <mat-card class="post-list">
 
-<mat-card class="votes">
- <button
- *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()"
- mat-flat-button color="primary" (click)="upVote()">
-   UpVote
-  </button>
-  {{post.upvoteCount - post.downvoteCount}}
-  <button
-  *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()"
-  mat-flat-button color="warn" (click)="downVote()">
-    DownVote
-  </button>
+    <!-- General Information -->
+    <p class="post-information">
+      {{post.category}}
+      {{post.createdAt |  date:'dd/MM/yyyy'}}
+    </p>
+    
+    <mat-card class="title">
+      {{post.title}}
+    </mat-card>
+    
+    <mat-card class="text">
+      {{post.text}}
+    </mat-card>
 
-</mat-card>
+    <!-- UpVote/DownVote -->
+    <mat-card class="votes">
+      <button
+        *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()"
+        mat-flat-button color="primary" (click)="upvotePost(post)">
+        UpVote
+      </button>
+       {{post.upvoteCount - post.downvoteCount}}
+      <button
+        *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()"
+        mat-flat-button color="warn" (click)="downvotePost(post)">
+        DownVote
+      </button>
+    </mat-card>
 
-  <button
-  *ngIf="userService.getLoggedIn() && !userService.getIsAdmin()"
-  mat-flat-button color="primary" (click)="openPopUp()">
-    Comment
-  </button>
-
-<button
-*ngIf="userService.getLoggedIn() && userService.getUser()?.userId === post.userId || userService.getIsAdmin()"
-mat-flat-button color="warn" (click)="delete()">
-  Delete Post
-</button>
-
-<button
-*ngIf="userService.getLoggedIn() && userService.getUser()?.userId === post.userId"
-mat-flat-button color="accent" (click)="openEditPopUp()">
-  Edit Post
-</button>
-
-  <mat-list>
-     <mat-list-item *ngFor="let comment of post.comments">
-      <mat-card>
-          {{comment.text}}
-        </mat-card>
-        <app-comment
-          [comment]="comment"
-          (deleteEvent)="deleteComment($event)"
-        ></app-comment>
-      </mat-list-item>
-    </mat-list>
-
-<div *ngIf="post.comments.length === 0">
- <mat-card>
-      No comments available.
-  </mat-card>
-</div>
-<!--
-  <mat-card>
-    Input field with list name
-    <mat-form-field>
-      <input matInput placeholder="List Name" [(ngModel)]="post.name" name="name" (ngModelChange)="updateList()">
-    </mat-form-field>
-    Button to delete list
-    <button mat-flat-button color="warn" (click)="deleteList()" [disabled]="post.comments.length != 0">
-      Delete list
+    <!-- Post Interactions --> 
+    <button
+      *ngIf="userService.getLoggedIn() && userService.getUser()?.userId === post.userId || userService.getIsAdmin()"
+      mat-flat-button color="warn" (click)="deletePost(post)">
+      Delete Post
+    </button>
+    
+    <button
+      *ngIf="userService.getLoggedIn() && userService.getUser()?.userId === post.userId"
+      mat-flat-button color="accent" (click)="openEditWindow(post)">
+      Edit Post
     </button>
 
-    <br>
+    <button
+      mat-flat-button color="primary" (click)="showComments(post, 'Comments')">
+      Show comments
+    </button>
 
-
- <mat-card>
- <hr>
-
-  List of CommentComponents displaying all items of list
-   <mat-list>
-      <mat-list-item *ngFor="let comment of post.comments">
-        <app-comment
-          [comment]="comment"
-          (update)="updateItem($event)"
-          (delete)="deleteItem($event)"
-        ></app-comment>
-      </mat-list-item>
-    </mat-list>
-
-    <div *ngIf="post.comments.length === 0">
-      No Posts available.
-    </div>
   </mat-card>
-
-    <Input field for new item name
-    <mat-form-field>
-      <input matInput placeholder="Item Name" [(ngModel)]="newCommentName">
-    </mat-form-field>
-    <Button to create item
-    <button mat-flat-button color="primary" (click)="createItem()" [disabled]="newCommentName === ''">Create new item</button>
--->
+</mat-list-item>
+<!--/////////////////////////////////////////////////////////////////////////////////////-->
