+++ ese2021-project-scaffolding/scaffolding/frontend/src/app/profile/profile.component.html
@@ -1,9 +1,9 @@
+  
+  <h3>Profile Page</h3>
   <!--Returns what was written in the register-->
   <!--TODO: Make it possible to change the information-->
 
   <!--
-
-  <h3>Profile Page</h3>
 
   <p>Username: {{user?.username}}</p>
