+++ ese2021-project-scaffolding/scaffolding/frontend/src/app/services/post-comment.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PostCommentService } from './post-comment.service';
+
+describe('UserService', () => {
+  let service: PostCommentService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PostCommentService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
