+++ ese2021-project-scaffolding/scaffolding/frontend/src/app/services/post-comment.service.ts
@@ -1,0 +1,51 @@
+import { Injectable } from '@angular/core';
+import { Post } from '../models/post.model';
+import { Subject } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class PostCommentService {
+
+  /*******************************************************************************************************************
+   * VARIABLES
+   ******************************************************************************************************************/
+
+  private post: Post = new Post(0, '', '', 0, 0, 0, 0, [], '', '');
+
+  /*******************************************************************************************************************
+   * OBSERVABLE SOURCES & STREAMS
+   ******************************************************************************************************************/
+
+  // Observable Sources
+  private postSource = new Subject<Post>();
+
+  // Observable Streams
+  post$ = this.postSource.asObservable();
+
+  /*******************************************************************************************************************
+   * GETTERS
+   ******************************************************************************************************************/
+
+  getPost(): Post {
+    return this.post;
+  }
+
+  /*******************************************************************************************************************
+   * SETTERS
+   ******************************************************************************************************************/
+
+  setPost(newPost: Post): void {
+    this.postSource.next(newPost);
+  }
+
+  /*******************************************************************************************************************
+   * CONSTRUCTOR
+   ******************************************************************************************************************/
+
+  constructor() {
+    // Observer
+    this.post$.subscribe(res => this.post = res);
+    this.setPost(new Post(0, '', '', 0, 0, 0, 0, [], '', ''));
+  }
+}
